ORA-19 fix CreateSeniorRequestMapper bug and amend Senior and Staff id columns (#9)

+++ backend/seniorsync/src/main/java/orangle/seniorsync/crm/requestmanagement/dto/CreateSeniorRequestDto.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record CreateSeniorRequestDto(
         @NotNull Long seniorId,
-        @NotNull Integer requestTypeId,
+        @NotNull Long requestTypeId,
         @NotBlank String title,
         @NotBlank String description,
         @NotNull Short priority
